+++ Cesium3DTilesSelection/test/TestTilesetContentManager.cpp
@@ -50,6 +50,7 @@
 #include <CesiumUtility/CreditSystem.h>
 #include <CesiumUtility/IntrusivePointer.h>
 #include <CesiumUtility/Math.h>
+#include <CesiumUtility/Uri.h>
 
 #include <doctest/doctest.h>
 #include <glm/common.hpp>
@@ -898,8 +899,9 @@ TEST_CASE("Test the tileset content manager's post processing for gltf") {
   SUBCASE("Resolve external buffers") {
     // create mock loader
     CesiumGltfReader::GltfReader gltfReader;
-    std::vector<std::byte> gltfBoxFile =
-        readFile(testDataPath / "gltf" / "box" / "Box.gltf");
+    std::filesystem::path boxPath = testDataPath / "gltf" / "box";
+    std::string fileName = (boxPath / "Box.gltf").string();
+    std::vector<std::byte> gltfBoxFile = readFile(fileName);
     auto modelReadResult = gltfReader.readGltf(gltfBoxFile);
 
     // check that this model has external buffer and it's not loaded
@@ -916,9 +918,37 @@ TEST_CASE("Test the tileset content manager's post processing for gltf") {
       CHECK(buffer.cesium.data.size() == 0);
     }
 
+    // add external buffer to the completed request
+    std::filesystem::path binPath = boxPath / "Box0.bin";
+    pMockedAssetAccessor->mockCompletedRequests.insert(
+        {Uri::nativePathToUriPath(binPath.string()),
+         createMockRequest(binPath)});
+
+    // Load the model and resolve external content
+    CesiumGltf::Model tileModel;
+    {
+      auto future =
+          gltfReader
+              .readGltfAndExternalData(
+                  gltfBoxFile,
+                  asyncSystem,
+                  CesiumAsync::HttpHeaders(),
+                  pMockedAssetAccessor,
+                  Uri::nativePathToUriPath(fileName))
+              .thenInMainThread(
+                  [&tileModel](
+                      CesiumGltfReader::GltfReaderResult&& externalReadResult) {
+                    CHECK(externalReadResult.errors.empty());
+                    CHECK(externalReadResult.warnings.empty());
+                    CHECK(externalReadResult.model);
+                    tileModel = std::move(*externalReadResult.model);
+                    return externalReadResult;
+                  });
+      future.waitInMainThread();
+    }
     auto pMockedLoader = std::make_unique<SimpleTilesetContentLoader>();
     pMockedLoader->mockLoadTileContent = {
-        std::move(*modelReadResult.model),
+        std::move(tileModel),
         CesiumGeometry::Axis::Y,
         std::nullopt,
         std::nullopt,
@@ -929,11 +959,6 @@ TEST_CASE("Test the tileset content manager's post processing for gltf") {
         TileLoadResultState::Success,
         Ellipsoid::WGS84};
     pMockedLoader->mockCreateTileChildren = {{}, TileLoadResultState::Failed};
-
-    // add external buffer to the completed request
-    pMockedAssetAccessor->mockCompletedRequests.insert(
-        {"Box0.bin",
-         createMockRequest(testDataPath / "gltf" / "box" / "Box0.bin")});
 
     // create tile
     auto pRootTile = std::make_unique<Tile>(pMockedLoader.get());